pyproject.toml
[tool.poetry]
name = "PythonMake"
version = "0.0.4"
description = "Makefile but modern"
authors = ["hoangson <lehoangsonsg@gmail.com>"]
readme = "README.md"
packages = [
    {include="PyMake"}
]

[tool.poetry.dependencies]
python = ">=3.10"
PyYAML = "^6.0.1"
pydantic = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-cov = "^4.1.0"
ruff = "^0.0.287"
black = "^23.9.0"
pre-commit = "^3.4.0"

[tool.poetry.scripts]
pymake = "PyMake.main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "@(abc\\.)?abstractmethod",
    "def __getattr__"
]

[tool.ruff]
line-length = 88
src = ["PyMake"]
target-version = "py310"
